Conflicts:
	TourPilot/bin/TourPilot.apk
	TourPilot/bin/classes.dex
	TourPilot/bin/classes/isebase/cognito/tourpilot/Activity/OptionsActivity.class
	TourPilot/bin/classes/isebase/cognito/tourpilot/Activity/ToursActivity.class
	TourPilot/bin/classes/isebase/cognito/tourpilot/Activity/WorkersActivity.class
	TourPilot/bin/classes/isebase/cognito/tourpilot/Data/BaseObject/BaseObjectManager.class
	TourPilot/bin/classes/isebase/cognito/tourpilot/DataBase/DataBaseWrapper.class
	TourPilot/bin/classes/isebase/cognito/tourpilot/R$menu.class
	TourPilot/bin/resources.ap_
	TourPilot/src/isebase/cognito/tourpilot/Activity/WorkersActivity.java
	TourPilot/src/isebase/cognito/tourpilot/Data/BaseObject/BaseObjectManager.java

+++ TourPilot/src/isebase/cognito/tourpilot/Data/BaseObject/BaseObjectManager.java
@@ -268,8 +268,8 @@ public abstract class BaseObjectManager<T> {
 				if (!method.getReturnType().equals(Void.TYPE))
 					continue;
 				MapField annos = method.getAnnotation(MapField.class);
-				if (annos == null)
-					continue;
+				if (annos == null) 
+					continue;			
 				if (method.getParameterTypes()[0].equals(int.class))
 					method.invoke(item, cursor.getInt(cursor
 							.getColumnIndex(annos.DatabaseField())));
@@ -295,10 +295,8 @@ public abstract class BaseObjectManager<T> {
 					method.invoke(item, cursor.getInt(cursor
 							.getColumnIndex(annos.DatabaseField())) == 1);
 				else if (method.getParameterTypes()[0].equals(Date.class))
-					method.invoke(
-							item,
-							new Date(cursor.getInt(cursor.getColumnIndex(annos
-									.DatabaseField()))));
+					method.invoke(item, new Date(cursor.getLong(cursor
+							.getColumnIndex(annos.DatabaseField()))));
 
 			} catch (Exception e) {
 				e.printStackTrace();
